style: Assert over frame counts without crossing signedness
reached_k_ is an int that starts at -1, but each of these asserts sits
past a guard that has already raised it, so the frame count it is
compared against is the same kind of quantity.

+++ engines/ic3base.cpp
@@ -436,7 +436,7 @@ ProverResult IC3Base::step(int i)
   // reached_k_ is the number of transitions that have been checked
   // at this point there are reached_k_ + 1 frames that don't
   // intersect bad, and reached_k_ + 2 frames overall
-  assert(reached_k_ == frontier_idx());
+  assert(static_cast<size_t>(reached_k_) == frontier_idx());
   logger.log(1, "Blocking phase at frame {}", i);
   if (!block_all()) {
     // counter-example
